[refactor] 메인페이지 종합 아파트 정보 및 공지사항 1차 컴포넌트화 #27
