+++ src/components/Abstraction/DropDown/DropDown.module.scss
@@ -6,10 +6,10 @@
     align-items: center;
     justify-content: space-between;
     &.minWidth {
-      width: 7rem;
+      width: 9rem;
     }
     &.maxWidth {
-      width: 10rem;
+      width: 12rem;
     }
     gap: 0.4rem;
     padding: 0.4rem 0rem;
@@ -20,14 +20,14 @@
     .selectContent {
       padding-left: 0.8rem;
       &.small {
-        font-size: 1rem;
+        font-size: 1.6rem;
       }
       &.medium {
-        font-size: 1.3rem;
+        font-size: 2rem;
       }
-      img {
-        margin-right: 1.2rem;
-      }
+    }
+    img {
+      width: 2.4rem;
     }
   }
   .dropDownBorder {
