+++ src/components/Abstraction/DropDown/DropDownList.module.scss
@@ -4,6 +4,7 @@
   border-bottom: none;
   padding: 0.4rem 2rem 0.4rem 0.8rem;
   padding-right: 0%;
+  font-size: 1.6rem;
   span {
     width: 100%;
   }
