+++ src/components/Common/ApartSearch.module.scss
@@ -1,0 +1,24 @@
+@use "@/styles/colors.scss";
+
+.apartSearch {
+  display: flex;
+  align-items: center;
+  max-width: 100%;
+  position: relative;
+  .apartSearchInput {
+    width: 100%;
+    border-top-left-radius: 10rem;
+    border-bottom-left-radius: 10rem;
+    padding: 1rem 0rem 1rem 2.4rem;
+    border: 0.1rem solid colors.$mainRed;
+    font-size: 1.6rem;
+    outline: none;
+    &::placeholder {
+      font-size: 1.6rem;
+    }
+  }
+  img {
+    position: absolute;
+    right: 1.6rem;
+  }
+}
