+++ src/components/Main/WholeApartInfo/CurrentApartsStatus.module.scss
@@ -1,0 +1,49 @@
+@use "@/styles/colors.scss";
+
+.currentApartStatus {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+  margin-bottom: 1.6rem;
+  //타이틀
+  .title {
+    background-color: colors.$mainNavy;
+    color: #f5f5f5;
+    font-size: 2rem;
+    font-weight: 600;
+    padding: 1rem 0rem 1rem 2.4rem;
+  }
+  //평균매매가
+  .averageSellingPrice {
+    display: flex;
+    .listTitle {
+      width: 50%;
+      display: flex;
+      flex-direction: column;
+      gap: 0.8rem;
+      justify-content: center;
+      background-color: colors.$mainTitle;
+      padding: 1.8rem 0rem 1.8rem 2.4rem;
+      .listMainTitle {
+        font-size: 1.6rem;
+        font-weight: 600;
+        color: #f5f5f5;
+      }
+      .listSubTitle {
+        font-size: 2rem;
+        color: #f5f5f5;
+      }
+    }
+    .listContent {
+      display: flex;
+      align-items: center;
+      flex-direction: row-reverse;
+      padding: 1.8rem 2.4rem 1.8rem 0rem;
+      width: 50%;
+      font-size: 2.2rem;
+      font-family: "Pretendard-Bold", sans-serif;
+      color: #f5f5f5;
+      background-color: #142337;
+    }
+  }
+}
