+++ src/components/Main/WholeApartInfo/ServiceNotice.module.scss
@@ -1,0 +1,80 @@
+@use "@/styles/colors.scss";
+
+//공지사항
+.serviceNotice {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  .serviceNoticeTitle {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    background-color: colors.$mainNavy;
+    color: #f5f5f5;
+    padding: 0.8rem 0.8rem 0.8rem 2.4rem;
+    margin-bottom: 1rem;
+    .title {
+      font-size: 2rem;
+      font-family: "Pretendard-Bold", sans-serif;
+    }
+  }
+  //공지사항 리스트 모음
+  .noticeLists {
+    display: flex;
+    flex-direction: column;
+    gap: 1.6rem;
+    height: 100%;
+    li {
+      border: 0.1rem solid #aaafb9;
+      padding: 2.4rem 3rem;
+      height: 33%;
+      .noticeListTitle {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        margin-bottom: 0.8rem;
+        strong {
+          font-size: 1.7rem;
+          font-weight: 700;
+          color: #142337;
+        }
+        span {
+          font-size: 1.5rem;
+          color: #aaafb9;
+        }
+      }
+      .noticeListContent {
+        font-size: 1.6rem;
+        line-height: 2rem;
+        color: #1e4164;
+        display: -webkit-box;
+        -webkit-line-clamp: 3;
+        -webkit-box-orient: vertical;
+        overflow: hidden;
+      }
+    }
+  }
+  .pagination {
+    margin-top: auto;
+    display: flex;
+    justify-content: center;
+    gap: 0.4rem;
+    padding-top: 2.7rem;
+    .pageContainer {
+      display: flex;
+      justify-content: center;
+      gap: 0.8rem;
+      .pageLink {
+        font-size: 1.4rem;
+        font-weight: 500;
+        line-height: 2rem;
+        color: colors.$mainTitle;
+      }
+      .pageActive {
+        font-weight: 700;
+        color: colors.$mainRed;
+      }
+    }
+  }
+}
