+++ src/components/Main/WholeApartInfo/WholeApartInfo.module.scss
@@ -1,157 +1,6 @@
-@use "@/styles/colors.scss";
-
 .homeRight {
   width: 45%;
   display: flex;
   flex-direction: column;
   gap: 1.6rem;
-  // 아파트 검색
-  .apartSearch {
-    display: flex;
-    align-items: center;
-    max-width: 100%;
-    position: relative;
-    .apartSearchInput {
-      width: 100%;
-      border-top-left-radius: 10rem;
-      border-bottom-left-radius: 10rem;
-      padding: 1rem 0rem 1rem 2.4rem;
-      border: 0.1rem solid colors.$mainRed;
-      font-size: 1.6rem;
-      outline: none;
-      &::placeholder {
-        font-size: 1.6rem;
-      }
-    }
-    img {
-      position: absolute;
-      right: 1.6rem;
-    }
-  }
-  //아파트 근황
-  .currentApartStatus {
-    display: flex;
-    flex-direction: column;
-    gap: 0.8rem;
-    margin-bottom: 1.6rem;
-    //타이틀
-    .title {
-      background-color: colors.$mainNavy;
-      color: #f5f5f5;
-      font-size: 2rem;
-      font-weight: 600;
-      padding: 1rem 0rem 1rem 2.4rem;
-    }
-    //평균매매가
-    .averageSellingPrice {
-      display: flex;
-      .listTitle {
-        width: 50%;
-        display: flex;
-        flex-direction: column;
-        gap: 0.8rem;
-        justify-content: center;
-        background-color: colors.$mainTitle;
-        padding: 1.8rem 0rem 1.8rem 2.4rem;
-        .listMainTitle {
-          font-size: 1.6rem;
-          font-weight: 600;
-          color: #f5f5f5;
-        }
-        .listSubTitle {
-          font-size: 2rem;
-          color: #f5f5f5;
-        }
-      }
-      .listContent {
-        display: flex;
-        align-items: center;
-        flex-direction: row-reverse;
-        padding: 1.8rem 2.4rem 1.8rem 0rem;
-        width: 50%;
-        font-size: 2.2rem;
-        font-family: "Pretendard-Bold", sans-serif;
-        color: #f5f5f5;
-        background-color: #142337;
-      }
-    }
-  }
-  //공지사항
-  .serviceNotice {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100%;
-    .serviceNoticeTitle {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      background-color: colors.$mainNavy;
-      color: #f5f5f5;
-      padding: 0.8rem 0.8rem 0.8rem 2.4rem;
-      margin-bottom: 1rem;
-      .title {
-        font-size: 2rem;
-        font-family: "Pretendard-Bold", sans-serif;
-      }
-    }
-    //공지사항 리스트 모음
-    .noticeLists {
-      display: flex;
-      flex-direction: column;
-      gap: 1.6rem;
-      height: 100%;
-      li {
-        border: 0.1rem solid #aaafb9;
-        padding: 2.4rem 3rem;
-        height: 33%;
-        .noticeListTitle {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          margin-bottom: 0.8rem;
-          strong {
-            font-size: 1.7rem;
-            font-weight: 700;
-            color: #142337;
-          }
-          span {
-            font-size: 1.5rem;
-            color: #aaafb9;
-          }
-        }
-        .noticeListContent {
-          font-size: 1.6rem;
-          line-height: 2rem;
-          color: #1e4164;
-          display: -webkit-box;
-          -webkit-line-clamp: 3;
-          -webkit-box-orient: vertical;
-          overflow: hidden;
-        }
-      }
-    }
-    .pagination {
-      margin-top: auto;
-      display: flex;
-      justify-content: center;
-      gap: 0.4rem;
-      padding-top: 2.7rem;
-      .pageContainer {
-        display: flex;
-        justify-content: center;
-        gap: 0.8rem;
-        .pageLink {
-          font-size: 1.4rem;
-          font-weight: 500;
-          line-height: 2rem;
-          color: colors.$mainTitle;
-        }
-        .pageActive {
-          font-weight: 700;
-          color: colors.$mainRed;
-        }
-      }
-    }
-  }
 }
