+++ src/pages/report.module.scss
@@ -5,19 +5,19 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  gap: 1.2rem;
+  gap: 2.4rem;
   border: 0.1rem solid #abb0b7;
   border-radius: 2rem;
   padding: 2.4rem;
   box-sizing: border-box;
   p {
-    font-size: 1.6rem;
+    font-size: 2.4rem;
     font-weight: 600;
     text-align: center;
     color: #435464;
   }
   a {
-    font-size: 2rem;
+    font-size: 2.8rem;
     font-weight: 600;
     color: #fff;
     padding: 1.2rem 5rem;
